# Optimizations for reading binary strings.