## Description of changes

This PR optimizes the process by which the binary `IonReader` reads `string`s.

### Performance change summary

#### Reading small strings
When decoding strings that are small enough to fit in the reusable decoding buffer:

&nbsp; | Time Required | GC Time
| ----------- | :--------------: | :---------: |
| Before | 615.351 ± 34.344 ms/op | 273ms | 
| After | 380.573 ± 11.136   ms/op | 185ms | 
| &nbsp; | 38.15% faster | 32.23% lower |

#### Reading big strings

&nbsp; | Time Required | GC Time
| ----------- | :--------------: | :---------: |
| Before | 867.368 ± 21.804 ms/op | 82ms | 
| After | 466.461 ± 21.796 ms/op | 54ms | 
| &nbsp; | 46.22% faster | 34.15% lower |

### Details

In order to read a `string` of length `n`, the old version of `IonBinaryReaderRawX#readString` will:
* Allocate and zero out a `char[]` of length `n` to store decoded `char`s.
* Read each byte from the data source individually, requiring `n` calls to `read()` and `n` end-of-file checks.
* Use our hand-written utf-8 decoder to turn the bytes being read into `char`s and eventually a `String`.

The new version created by this PR will:
* Use a reusable `CharBuffer` to store decoded `char`s if the `string` being read is small enough.
* Read the `string` into a reusable buffer in pages of up to 4096 bytes, requiring `n/4096` calls to `read(buffer)` and `n/4096` end-of-file checks. For most strings, only a single read+check will be required.
* Use Java's standard `CharsetDecoder` class to perform utf-8 decoding.

### Benchmarks \[[source](https://gist.github.com/zslayton/dedabd7078c3a7e4ab96cfb55b36f594)\]

#### Small Strings

This test creates 2.5 million strings that are **small enough to fit in the reader's reusable decoding buffer** and writes them to a `byte []` as binary Ion. The timed portion of the benchmark reads each string in the stream.

**before**
```
Benchmark                                Mode  Cnt          Score          Error   Units
read                                       ss   20        615.351 ±       34.344   ms/op
read:Heap usage                            ss   20        889.843 ±       20.990      MB
read:·gc.churn.{PS Eden Space}             ss   20        867.668 ±       55.462  MB/sec
read:·gc.churn.{PS Eden Space}.norm        ss   20  909816332.578 ± 58156173.178    B/op
read:·gc.churn.{PS Old Gen}                ss   20            ≈ 0                 MB/sec
read:·gc.churn.{PS Old Gen}.norm           ss   20            ≈ 0                   B/op
read:·gc.churn.{PS Survivor Space}         ss   20          0.028 ±        0.023  MB/sec
read:·gc.churn.{PS Survivor Space}.norm    ss   20      29597.107 ±    23865.736    B/op
read:·gc.count                             ss   20         60.000                 counts
read:·gc.time                              ss   20        273.000                     ms
```

**after**
```
Benchmark                                Mode  Cnt          Score          Error   Units
read                                       ss   20        380.573 ±       11.136   ms/op
read:Heap usage                            ss   20        863.260 ±        9.357      MB
read:·gc.churn.{PS Eden Space}             ss   20        684.124 ±       35.296  MB/sec
read:·gc.churn.{PS Eden Space}.norm        ss   20  717356058.885 ± 37010139.278    B/op
read:·gc.churn.{PS Old Gen}                ss   20            ≈ 0                 MB/sec
read:·gc.churn.{PS Old Gen}.norm           ss   20            ≈ 0                   B/op
read:·gc.churn.{PS Survivor Space}         ss   20          0.008 ±        0.022  MB/sec
read:·gc.churn.{PS Survivor Space}.norm    ss   20       8749.936 ±    23386.456    B/op
read:·gc.count                             ss   20         40.000                 counts
read:·gc.time                              ss   20        185.000                     ms
```

#### Big Strings


This test creates 20,000 strings that are **too big to fit in the reader's reusable decoding buffer** and writes them to a `byte []` as binary Ion. The timed portion of the benchmark reads each string in the stream.

**before**
```
Benchmark                                Mode  Cnt          Score          Error   Units
read                                       ss   20        867.368 ±       21.804   ms/op
read:Heap usage                            ss   20       1143.081 ±       11.483      MB
read:·gc.churn.{PS Eden Space}             ss   20        544.001 ±       16.853  MB/sec
read:·gc.churn.{PS Eden Space}.norm        ss   20  570426378.431 ± 17671260.123    B/op
read:·gc.churn.{PS Old Gen}                ss   20            ≈ 0                 MB/sec
read:·gc.churn.{PS Old Gen}.norm           ss   20            ≈ 0                   B/op
read:·gc.churn.{PS Survivor Space}         ss   20            ≈ 0                 MB/sec
read:·gc.churn.{PS Survivor Space}.norm    ss   20            ≈ 0                   B/op
read:·gc.count                             ss   20         20.000                 counts
read:·gc.time                              ss   20         82.000                     ms
```

**after**
```
Benchmark                                Mode  Cnt          Score           Error   Units
read                                       ss   20        466.461 ±        21.796   ms/op
read:Heap usage                            ss   20       1396.696 ±       217.578      MB
read:·gc.churn.{PS Eden Space}             ss   20        483.291 ±       432.141  MB/sec
read:·gc.churn.{PS Eden Space}.norm        ss   20  506766941.853 ± 453132888.558    B/op
read:·gc.churn.{PS Old Gen}                ss   20            ≈ 0                  MB/sec
read:·gc.churn.{PS Old Gen}.norm           ss   20            ≈ 0                    B/op
read:·gc.churn.{PS Survivor Space}         ss   20            ≈ 0                  MB/sec
read:·gc.churn.{PS Survivor Space}.norm    ss   20            ≈ 0                    B/op
read:·gc.count                             ss   20         10.000                  counts
read:·gc.time                              ss   20         54.000                      ms
```

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.