@zslayton 2019-11-21
> Nice change. Does the benchmark include both small strings (that hit the `readStringWithReusableBuffer` path), and large strings that hit the fallback path? It would be nice to also separate the two cases to confirm that both paths improved.

I've amended the PR description to break out the results by string size. I've also included a high-level summary of the change for each in tables at the top.